Restructure and fix stubgen integration tests
