--- wasm-rpc-stubgen/tests-integration/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "golem-wasm-rpc-stubgen-tests-integration"
-version = "0.0.0"
-edition = "2021"
-license = "Apache-2.0"
-homepage = "https://golem.cloud"
-repository = "https://github.com/golemcloud/golem-cli"
-description = "Golem WASM RPC stub generator - integration tests"
-
-[lib]
-path = "src/lib.rs"
-harness = false
-
-[[test]]
-name = "app"
-harness = false
-
-[[test]]
-name = "compose"
-harness = false
-
-[[test]]
-name = "stub_wasm"
-harness = false
-
-[dependencies]
-golem-wasm-rpc-stubgen = { path = "../../wasm-rpc-stubgen", version = "=0.0.0" }
-
-golem-wasm-ast = { workspace = true }
-
-assert2 = { workspace = true }
-fs_extra = { workspace = true }
-tempfile = { workspace = true }
-tokio = { workspace = true }
-
-[dev-dependencies]
-test-r = { workspace = true }
